Repair add some feaure now all running

+++ src/Controllers/user.controller.ts
@@ -3,12 +3,13 @@ import { UserService } from "../Services/user.service";
 import { inject } from "inversify";
 import { TYPES } from "../Types/types"
 import { Iuser } from "../Interfaces/model.interface";
-import { ApiError } from "../Utiles/Apierror";
+import { ApiError } from "../Utiles";
 import { statuscode } from "../Constans/stacode";
 import { errMSG } from "../Constans/message";
 import { Request, Response } from "express";
-import { IupdateUser } from "../Interfaces/request.interface";
 import { upload } from "../Middlewares/multer.midddleware";
+import { Auth } from "../Middlewares/auth.middleware";
+import { Role } from "../Middlewares/role.middleware";
 
 @controller('/user')
 export class UserController {
@@ -29,7 +30,6 @@ export class UserController {
       const signupData: Iuser = req.body as unknown as Iuser;
 
       const files = req.files as { [fieldname: string]: Express.Multer.File[] };
-
       const profilePictureLocalpath = files?.profilePicture?.[0]?.path;
       signupData.profilepic = profilePictureLocalpath;
 
@@ -37,7 +37,7 @@ export class UserController {
 
       res.status(created_user.statuscode).json(created_user);
     } catch (error: any) {
-      res.status(error.stacode || 500).json({ message: error.message || errMSG.InternalServerErrorResult })
+      res.status(error.statuscode || statuscode.INTERNALSERVERERROR).json({ message: error.message || errMSG.InternalServerErrorResult })
     }
   }
 
@@ -47,59 +47,70 @@ export class UserController {
       const loginData: Iuser = req.body as unknown as Iuser;
 
       if ([loginData.email, loginData.password].some((field) => field.trim() == "")) {
-        throw new ApiError(statuscode.NotAcceptable, errMSG.exsistuser);
+        throw new ApiError(statuscode.NOTACCEPTABLE, errMSG.exsistuser);
       }
 
       const login_user = await this.user.loginUser(loginData);
 
       res.status(login_user.statuscode).json(login_user);
     } catch (error: any) {
-      res.status(error.stacode).json({ message: error.message || errMSG.InternalServerErrorResult })
+      res.status(error.statuscode || statuscode.INTERNALSERVERERROR).json({ message: error.message || errMSG.InternalServerErrorResult })
     }
 
   }
 
-  @httpDelete('/delete')
+  @httpDelete('/delete', Auth, Role)
   async delete(req: Request, res: Response) {
     try {
       const userId = req.body.userId;
       if (!userId) {
-        throw new ApiError(statuscode.NotAcceptable, errMSG.exsistuser);
+        throw new ApiError(statuscode.NOTACCEPTABLE, errMSG.exsistuser);
+      }
+      if (req.body.ROLE === 'admin' && userId != req.body.USERID) {
+        throw new ApiError(statuscode.NOTACCEPTABLE, errMSG.notValidRole(req.body.ROLE, 'update the another user'));
       }
 
       const deleted_user = await this.user.deleteUser(userId);
 
       res.status(deleted_user.statuscode).json(deleted_user);
     } catch (error) {
-      res.status(error.stacode).json({ message: error.message || errMSG.InternalServerErrorResult })
+      res.status(error.statuscode || statuscode.INTERNALSERVERERROR).json({ message: error.message || errMSG.InternalServerErrorResult })
     }
   }
 
-  @httpPut('/update')
-  async update(req: Request, res: Response) {
+  @httpPut('/update', Auth, Role, upload.fields([{
+    name: 'profilePicture',
+    maxCount: 1
+  }]))
+  async update(req: Request, res: Response,) {
     try {
-      const updateData: IupdateUser = req.body as unknown as IupdateUser;
+      const updateData: Iuser = req.body as unknown as Iuser;
 
-      if ([updateData.name, updateData.email, updateData.password, updateData.role].some((field) => field.trim() == "")) {
-        throw new ApiError(statuscode.NotAcceptable, errMSG.exsistuser);
+      if (updateData._id != req.body.USERID) {
+        throw new ApiError(statuscode.NOTACCEPTABLE, errMSG.notValidRole(req.body.ROLE, 'update the another user'));
       }
 
-      const updated_user = await this.user.updateUser(updateData);
+      const files = req.files as { [fieldname: string]: Express.Multer.File[] };
+      const profilePictureLocalpath = files?.profilePicture?.[0]?.path;
+      updateData.profilepic = profilePictureLocalpath;
+
+      const updated_user = profilePictureLocalpath ? await this.user.updateUserWithProfilePicture(updateData) : await this.user.updateUserWithoutProfilePicture(updateData)
 
       res.status(updated_user.statuscode).json(updated_user);
     } catch (error) {
-      res.status(error.stacode).json({ message: error.message || errMSG.InternalServerErrorResult })
+      res.status(error.statuscode || statuscode.INTERNALSERVERERROR).json({ message: error.message || errMSG.InternalServerErrorResult })
     }
   }
 
-  @httpGet('/getAll')
+  @httpGet('/getById/:id?', Auth, Role)
   async getAll(req: Request, res: Response) {
     try {
-      const allUser = await this.user.getAlluser();
+      const userId = req.params.id
+      const allUser = await this.user.getUserById(userId);
 
       res.status(allUser.statuscode).json(allUser)
     } catch (error) {
-      res.status(error.stacode).json({ message: error.message || errMSG.InternalServerErrorResult })
+      res.status(error.statuscode || statuscode.INTERNALSERVERERROR).json({ message: error.message || errMSG.InternalServerErrorResult })
     }
   }
 }
